requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
jinja2==3.1.6
python-dotenv==1.1.0
azure-identity==1.23.0
openai==1.78.1
aiohttp==3.11.18
python-multipart==0.0.20
pydantic==2.11.4
pydantic-settings==2.2.1
rich==14.0.0
azure-storage-blob==12.26.0
dotenv==0.9.9
azure-search-documents==11.5.3
